requirements.txt.txt
gym==0.26.2
stable-baselines3==1.8.0
matplotlib==3.7.1
imageio==2.31.1
PyOpenGL==3.1.7
numpy==1.24.3